% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bodleian.R
\name{bodleian}
\alias{bodleian}
\alias{bodsearch}
\alias{solosearch}
\alias{solo}
\title{Search the Oxford SOLO Bodleian Libraries catalogue}
\usage{
bodleian(search_terms)

bodsearch(search_terms)

solosearch(search_terms)

solo(search_terms)
}
\arguments{
\item{search_terms}{Search terms encapsulated in " ".}
}
\description{
Search the Bodleian Libraries catalogue directly from the R console without having to switching to the browser and opening a new tab first.
}
\examples{
bodleian("a book")
bodsearch("an author")
solosearch("a more interesting book")
solo("yet another book")
}
\keyword{Oxford}
\keyword{internet}
\keyword{library}
\keyword{workflow}
